pipeline improvement

+++ projects/mmdiff/models/utils.py
@@ -68,19 +68,19 @@ def get_beta_schedule(beta_schedule, *, beta_start, beta_end, num_diffusion_time
                 beta_start ** 0.5,
                 beta_end ** 0.5,
                 num_diffusion_timesteps,
-                dtype=np.float64,
+                dtype=np.float32,
             )
             ** 2
         )
     elif beta_schedule == "linear":
         betas = np.linspace(
-            beta_start, beta_end, num_diffusion_timesteps, dtype=np.float64
+            beta_start, beta_end, num_diffusion_timesteps, dtype=np.float32
         )
     elif beta_schedule == "const":
-        betas = beta_end * np.ones(num_diffusion_timesteps, dtype=np.float64)
+        betas = beta_end * np.ones(num_diffusion_timesteps, dtype=np.float32)
     elif beta_schedule == "jsd":  # 1/T, 1/(T-1), 1/(T-2), ..., 1
         betas = 1.0 / np.linspace(
-            num_diffusion_timesteps, 1, num_diffusion_timesteps, dtype=np.float64
+            num_diffusion_timesteps, 1, num_diffusion_timesteps, dtype=np.float32
         )
     elif beta_schedule == "sigmoid":
         betas = np.linspace(-6, 6, num_diffusion_timesteps)
